+++ transektcount/src/main/java/com/wmstein/transektcount/widgets/CountingWidgetLhInt.kt
@@ -1,12 +1,18 @@
 /*
- * Copyright (c) 2016 - 2023. Wilhelm Stein, Bonn, Germany.
+ * Copyright (c) 2016 - 2025. Wilhelm Stein, Bonn, Germany.
  */
 package com.wmstein.transektcount.widgets
 
 import android.content.Context
+import android.content.Context.WINDOW_SERVICE
+import android.graphics.Point
+import android.os.Build
 import android.util.AttributeSet
 import android.view.LayoutInflater
+import android.view.ViewGroup
+import android.view.WindowManager
 import android.widget.ImageButton
+import android.widget.LinearLayout
 import android.widget.RelativeLayout
 import android.widget.TextView
 import com.wmstein.transektcount.AutoFitText
@@ -19,16 +25,26 @@ import java.util.Objects
  * Created by wmstein on 06.09.2016
  * Last edited in Java on 2021-01-26,
  * converted to Kotlin on 2023-06-26,
- * last edited on 2023-10-06
+ * last edited on 2025-04-15
  */
 class CountingWidgetLhInt(context: Context, attrs: AttributeSet?) : RelativeLayout(context, attrs) {
+    var sHeight: Int = 0
+    var ht: Int = 0
+
     private val namef1i: TextView
     private val namef2i: TextView
     private val namef3i: TextView
     private val namepi: TextView
     private val nameli: TextView
     private val nameei: TextView
-    
+
+    private var idLHf1i: LinearLayout
+    private var idLHf2i: LinearLayout
+    private var idLHf3i: LinearLayout
+    private var idLHpi: LinearLayout
+    private var idLHli: LinearLayout
+    private var idLHei: LinearLayout
+
     // section internal counters
     private val countCountf1i: AutoFitText
     private val countCountf2i: AutoFitText
@@ -36,18 +52,27 @@ class CountingWidgetLhInt(context: Context, attrs: AttributeSet?) : RelativeLayo
     private val countCountpi: AutoFitText
     private val countCountli: AutoFitText
     private val countCountei: AutoFitText
+
     @JvmField
     var count: Count? = null
 
     init {
         val inflater = context.getSystemService(Context.LAYOUT_INFLATER_SERVICE) as LayoutInflater
         Objects.requireNonNull(inflater).inflate(R.layout.widget_counting_lhi, this, true)
+        idLHf1i = findViewById(R.id.idLHf1i)
+        idLHf2i = findViewById(R.id.idLHf2i)
+        idLHf3i = findViewById(R.id.idLHf3i)
+        idLHpi = findViewById(R.id.idLHpi)
+        idLHli = findViewById(R.id.idLHli)
+        idLHei = findViewById(R.id.idLHei)
+
         namef1i = findViewById(R.id.f1iNameLH)
         namef2i = findViewById(R.id.f2iNameLH)
         namef3i = findViewById(R.id.f3iNameLH)
         namepi = findViewById(R.id.piNameLH)
         nameli = findViewById(R.id.liNameLH)
         nameei = findViewById(R.id.eiNameLH)
+
         countCountf1i = findViewById(R.id.countCountLHf1i)
         countCountf2i = findViewById(R.id.countCountLHf2i)
         countCountf3i = findViewById(R.id.countCountLHf3i)
@@ -56,7 +81,52 @@ class CountingWidgetLhInt(context: Context, attrs: AttributeSet?) : RelativeLayo
         countCountei = findViewById(R.id.countCountLHei)
     }
 
+    @Suppress("RemoveRedundantQualifierName")
     fun setCountLHi(newcount: Count?) {
+        // Get screen size to adapt the counting view
+        val wm = checkNotNull(context.getSystemService(WINDOW_SERVICE) as WindowManager)
+        if (Build.VERSION.SDK_INT >= 30) {
+            val metrics = wm.currentWindowMetrics
+            sHeight = metrics.bounds.top + metrics.bounds.bottom
+        } else {
+            @Suppress("DEPRECATION")
+            val display = wm.defaultDisplay // deprecated in 30
+            val size = Point()
+            @Suppress("DEPRECATION")
+            display.getSize(size) // deprecated in 30
+            sHeight = size.y
+        }
+
+        // 4 different screen sizes
+        if (sHeight >= 2850)
+            ht = 140 // 125 -> 5 mm, 500 -> 24 mm, 50 -> 2 mm (in xml 38dp -> 4,5 mm)
+        else if (sHeight >= 2400)
+            ht = 125
+        else if (sHeight >= 2000)
+            ht = 110
+        else
+            ht = 94
+
+        val lparamsf1i: ViewGroup.LayoutParams = idLHf1i.layoutParams
+        lparamsf1i.height = ht
+        val lparamsf2i: ViewGroup.LayoutParams = idLHf2i.layoutParams
+        lparamsf2i.height = ht
+        val lparamsf3i: ViewGroup.LayoutParams = idLHf3i.layoutParams
+        lparamsf3i.height = ht
+        val lparamspi: ViewGroup.LayoutParams = idLHpi.layoutParams
+        lparamspi.height = ht
+        val lparamsli: ViewGroup.LayoutParams = idLHli.layoutParams
+        lparamsli.height = ht
+        val lparamsei: ViewGroup.LayoutParams = idLHei.layoutParams
+        lparamsei.height = ht
+
+        idLHf1i.layoutParams = lparamsf1i
+        idLHf2i.layoutParams = lparamsf2i
+        idLHf3i.layoutParams = lparamsf3i
+        idLHpi.layoutParams = lparamspi
+        idLHli.layoutParams = lparamsli
+        idLHei.layoutParams = lparamsei
+
         count = newcount
         namef1i.text = context.getString(R.string.countImagomfHint)
         namef2i.text = context.getString(R.string.countImagomHint)
@@ -64,12 +134,14 @@ class CountingWidgetLhInt(context: Context, attrs: AttributeSet?) : RelativeLayo
         namepi.text = context.getString(R.string.countPupaHint)
         nameli.text = context.getString(R.string.countLarvaHint)
         nameei.text = context.getString(R.string.countOvoHint)
+
         countCountf1i.text = count!!.count_f1i.toString()
         countCountf2i.text = count!!.count_f2i.toString()
         countCountf3i.text = count!!.count_f3i.toString()
         countCountpi.text = count!!.count_pi.toString()
         countCountli.text = count!!.count_li.toString()
         countCountei.text = count!!.count_ei.toString()
+
         val countUpf1eButton = findViewById<ImageButton>(R.id.buttonUpLHf1i)
         countUpf1eButton.tag = count!!.id
         val countUpf2eButton = findViewById<ImageButton>(R.id.buttonUpLHf2i)